requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
PyJWT==2.8.0
gunicorn==21.2.0
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
bleach==6.1.0
tencentcloud-sdk-python==3.0.1310
APScheduler==3.10.4
qcloud_cos_python==5.1.9.24
